requirements.txt
anthropic>=0.39.0
PyMuPDF>=1.24.0
pandas>=2.1.0
openpyxl>=3.1.0
pydantic>=2.5.0
python-dotenv>=1.0.0
flask>=3.0.0
